+++ temporaldata/temporaldata/data.py
@@ -52,7 +52,6 @@ class Data(object):
         ...     lfp=RegularTimeSeries(
         ...         raw=np.zeros((1000, 3)),
         ...         sampling_rate=250.,
-        ...         domain=Interval(0., 4.),
         ...     ),
         ...     units=ArrayDict(
         ...         id=np.array(["unit_0", "unit_1", "unit_2"]),